+++ src/AutoMapper.Extensions.ExpressionMapping/Properties/Resources.resx
@@ -102,8 +102,8 @@
     <value>Cannot create a binary expression for the following pair.  Node: {0}, Type: {1} and Node: {2}, Type: {3}.</value>
     <comment>0=leftNode; 1=leftNodeType; 2=rightNode; 3=rightNodeType</comment>
   </data>
-	<data name="cantRemapExpression" xml:space="preserve">
-    <value>Can't rempa expression</value>
+	<data name="cantMapExpression" xml:space="preserve">
+    <value>Can't map expression</value>
   </data>
 	<data name="expressionMapValueTypeMustMatchFormat" xml:space="preserve">
     <value>The source and destination types must be the same for expression mapping between literal types. Source Type: {0}, Source Description: {1}, Destination Type: {2}, Destination Property: {3}.</value>
@@ -144,7 +144,11 @@
     <comment>0=typeSource, 1=typeDestination; 2=sourceFullName</comment>
   </data>
 	<data name="argumentMustBeDelegateFormat" xml:space="preserve">
-    <value>Argument {0} must be a deledate type.</value>
+    <value>Argument {0} must be a delegate type.</value>
     <comment>0=argumentType</comment>
+  </data>
+	<data name="invalidTypeMappingsFormat" xml:space="preserve">
+    <value>Invalis type mappings. Source Type: {0}, Destination Type: {1}.</value>
+    <comment>0=sourceType; 1=destinationTyp;</comment>
   </data>
 </root>
